requirements.txt
jinja2
jsonschema
pytest
google-api-python-client
google-auth-oauthlib